refactor:
OrderGateway 삭제 및 response 위치 수정
